[#11550] test(idp-basic): Move BasicAuthOperationsIT from client-java
Relocate the built-in IDP Basic auth IT beside IdpRESTApiIT and remove the
idp-basic test dependency from client-java. Trim built-in IDP Web UI docs.
