+++ clients/client-java/build.gradle.kts
@@ -44,7 +44,6 @@ dependencies {
 
   testImplementation(project(":core"))
   testImplementation(project(":integration-test-common", "testArtifacts"))
-  testImplementation(project(":plugins:idp-basic"))
   testImplementation(project(":server"))
   testImplementation(project(":server-common"))
